# services/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/pacientes/{id}": {
      "get": {
        "operationId": "PacientesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "PacientesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePacienteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "PacientesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/pacientes/{pacienteId}/prontuarios": {
      "post": {
        "operationId": "ProntuariosController_create",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProntuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProntuarioDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "ProntuariosController_findAll",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProntuarioDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pacientes/{pacienteId}/prontuarios/{id}": {
      "get": {
        "operationId": "ProntuariosController_findOne",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProntuarioDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "ProntuariosController_update",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProntuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProntuarioDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "ProntuariosController_remove",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/pacientes/{pacienteId}/peso": {
      "get": {
        "operationId": "RegistrosController_listarPeso",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistroPesoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pacientes/{pacienteId}/diario": {
      "get": {
        "operationId": "RegistrosController_listarDiario",
        "parameters": [
          {
            "name": "pacienteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistroDiarioDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "NutriDeby API Gateway",
    "description": "Contrato do api-gateway. Fonte única de tipos para admin-web e pwa-patient — não editar à mão nos clientes.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePacienteDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProntuarioDto": {
        "type": "object",
        "properties": {
          "queixaPrincipal": {
            "type": "string"
          },
          "historicoClinico": {
            "type": "string"
          },
          "historicoFamiliar": {
            "type": "string"
          },
          "habitosAlimentares": {
            "type": "string"
          },
          "usoMedicamentos": {
            "type": "string"
          },
          "alergias": {
            "type": "string"
          },
          "intolerancias": {
            "type": "string"
          },
          "nivelAtividadeFisica": {
            "type": "string"
          },
          "observacoesGerais": {
            "type": "string"
          },
          "pesoKg": {
            "type": "number"
          },
          "alturaCm": {
            "type": "number"
          },
          "circunferenciaCintura": {
            "type": "number"
          },
          "circunferenciaQuadril": {
            "type": "number"
          },
          "percentualGordura": {
            "type": "number"
          },
          "imc": {
            "type": "number"
          }
        }
      },
      "ProntuarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pacienteId": {
            "type": "string"
          },
          "queixaPrincipal": {
            "type": "string",
            "nullable": true
          },
          "historicoClinico": {
            "type": "string",
            "nullable": true
          },
          "historicoFamiliar": {
            "type": "string",
            "nullable": true
          },
          "habitosAlimentares": {
            "type": "string",
            "nullable": true
          },
          "usoMedicamentos": {
            "type": "string",
            "nullable": true
          },
          "alergias": {
            "type": "string",
            "nullable": true
          },
          "intolerancias": {
            "type": "string",
            "nullable": true
          },
          "nivelAtividadeFisica": {
            "type": "string",
            "nullable": true
          },
          "observacoesGerais": {
            "type": "string",
            "nullable": true
          },
          "pesoKg": {
            "type": "string",
            "nullable": true,
            "description": "Decimal serializado como string, ex.: \"90.00\"."
          },
          "alturaCm": {
            "type": "string",
            "nullable": true
          },
          "circunferenciaCintura": {
            "type": "string",
            "nullable": true
          },
          "circunferenciaQuadril": {
            "type": "string",
            "nullable": true
          },
          "percentualGordura": {
            "type": "string",
            "nullable": true
          },
          "imc": {
            "type": "string",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "atualizadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "pacienteId",
          "queixaPrincipal",
          "historicoClinico",
          "historicoFamiliar",
          "habitosAlimentares",
          "usoMedicamentos",
          "alergias",
          "intolerancias",
          "nivelAtividadeFisica",
          "observacoesGerais",
          "pesoKg",
          "alturaCm",
          "circunferenciaCintura",
          "circunferenciaQuadril",
          "percentualGordura",
          "imc",
          "criadoEm",
          "atualizadoEm"
        ]
      },
      "UpdateProntuarioDto": {
        "type": "object",
        "properties": {}
      },
...
      "RegistroPesoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pacienteId": {
            "type": "string"
          },
          "pesoKg": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "registradoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "pacienteId",
          "pesoKg",
          "registradoEm"
        ]
      },
      "RegistroDiarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pacienteId": {
            "type": "string"
          },
          "texto": {
            "type": "string",
            "nullable": true
          },
          "fotoUrl": {
            "type": "string",
            "nullable": true
          },
          "registradoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "pacienteId",
          "texto",
          "fotoUrl",
          "registradoEm"
        ]
      },
...
    }
  }
}